Reduce most logging to 'info' level by default
The logging done by the init script is mainly used for testing
purposes and not needed by consumers.

In cases where this additional logging is needed, e.g., for debugging
injection in an environment, the property
`develocity.injection.debug=true` can be used to enable it.
